Update prune media command to delete in batches

+++ main/management/commands/prunemedia.py
@@ -1,9 +1,20 @@
 from django.core.management.base import BaseCommand
 from main.models import Media
+import logging
+
+logger = logging.getLogger(__name__)
 
 class Command(BaseCommand):
     help = 'Deletes any media files marked for deletion with null project.'
+    BATCH_SIZE = 100
 
     def handle(self, **options):
-        media = Media.objects.filter(project__isnull=True)
-        media.delete()
+        num_deleted = 0
+        while True:
+            media = Media.objects.filter(project__isnull=True)[:BATCH_SIZE]
+            num_media = media.count()
+            if num_media == 0:
+                break
+            media.delete()
+            num_deleted += num_media
+            logger.info(f"Deleted a total of {num_deleted} media...")
